refactor: code restructure needed for multiaccount
