+++ assets/copy.md
@@ -1,0 +1,46 @@
+# Store Copy
+
+
+## Google Play Store
+
+### Short Description
+A privacy focused matrix chat client
+
+
+### Description
+Syphon is privacy focused end-to-end encrypted capable matrix client aiming to feel comfortable and easy to use.
+
+That means:
+- end-to-end encryption for all available chats
+- no analytics, tracking, or telemetry
+- no third party APIs (Firebase, Facebook SDK, etc)
+- opt-in only for all signs of presence (typing bubbles, read receipts, etc) 
+- and much more
+
+E2EE is working for direct chats and group chats, but until we have a third party security audit, we cannot recommend using Syphon where proven and independently verified security is required.
+
+There's still a lot of work to be done for user experience, performance, and feature parity with everything in the Matrix ecosystem. We'll always be improving!
+
+We’re hoping that you can see the vision of the app! Want to help out or send us feedback? Feel free to email us at hello@syphon.org
+
+## App Store
+
+### Short Description
+
+### Description
+Syphon is privacy focused end-to-end encrypted capable matrix client aiming to feel comfortable and easy to use.
+
+That means:
+- end-to-end encryption for all available chats
+- no analytics, tracking, or telemetry
+- no third party APIs (Firebase, Facebook SDK, etc)
+- opt-in only for all signs of presence (typing bubbles, read receipts, etc) 
+- and much more
+
+E2EE is working for direct chats and group chats, but until we have a third party security audit, we cannot recommend using Syphon where proven and independently verified security is required.
+
+We know there's always more work to be done on user experience, performance, and feature parity with everything in the Matrix ecosystem. We're constantly improving!
+
+We’re hoping that you can see the vision of the app! Want to help out or send us feedback? Feel free to email us at hello@syphon.org
+
+## FDroid
