+++ lib/views/prelock.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/widgets.dart';
+
+class Prelock extends StatefulWidget {
+  @override
+  _PrelockState createState() => _PrelockState();
+}
+
+class _PrelockState extends State<Prelock> {
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    throw UnimplementedError();
+  }
+}
