+++ lib/views/syphon.dart
@@ -1,0 +1,211 @@
+import 'dart:async';
+
+import 'package:easy_localization/easy_localization.dart' as localization;
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
+
+import 'package:flutter_redux/flutter_redux.dart';
+import 'package:redux/redux.dart';
+import 'package:sembast/sembast.dart';
+import 'package:syphon/cache/index.dart';
+import 'package:syphon/global/formatters.dart';
+import 'package:syphon/global/notifications.dart';
+
+import 'package:syphon/global/themes.dart';
+import 'package:syphon/store/alerts/actions.dart';
+import 'package:syphon/store/auth/actions.dart';
+import 'package:syphon/store/events/messages/actions.dart';
+import 'package:syphon/store/index.dart';
+import 'package:syphon/store/settings/theme-settings/model.dart';
+import 'package:syphon/store/sync/actions.dart';
+import 'package:syphon/store/sync/background/storage.dart';
+import 'package:syphon/views/home/home-screen.dart';
+import 'package:syphon/views/intro/intro-screen.dart';
+import 'package:syphon/views/navigation.dart';
+
+class Syphon extends StatefulWidget {
+  final Database? cache;
+  final Database? storage;
+  final Store<AppState> store;
+
+  const Syphon(
+    this.store,
+    this.cache,
+    this.storage,
+  );
+
+  @override
+  SyphonState createState() => SyphonState(
+        store,
+        cache,
+        storage,
+      );
+}
+
+class SyphonState extends State<Syphon> with WidgetsBindingObserver {
+  final Database? cache;
+  final Database? storage;
+  final Store<AppState> store;
+  final GlobalKey<ScaffoldState> globalScaffold = GlobalKey<ScaffoldState>();
+
+  Widget defaultHome = HomeScreen();
+  StreamSubscription? alertsListener;
+
+  SyphonState(
+    this.store,
+    this.cache,
+    this.storage,
+  );
+
+  @override
+  void initState() {
+    WidgetsBinding.instance?.addObserver(this);
+    super.initState();
+
+    store.dispatch(initDeepLinks());
+    store.dispatch(initClientSecret());
+    store.dispatch(startAuthObserver());
+    store.dispatch(startAlertsObserver());
+
+    // init current auth state with current user
+    store.state.authStore.authObserver!.add(
+      store.state.authStore.user,
+    );
+
+    // mutate messages
+    store.dispatch(mutateMessagesAll());
+
+    final currentUser = store.state.authStore.user;
+    final authed = currentUser.accessToken != null;
+
+    if (!authed) {
+      defaultHome = IntroScreen();
+    }
+  }
+
+  @override
+  // ignore: avoid_void_async
+  void didChangeAppLifecycleState(AppLifecycleState state) async {
+    switch (state) {
+      case AppLifecycleState.resumed:
+        setupTheme(store.state.settingsStore.themeSettings);
+
+        dismissAllNotifications(
+          pluginInstance: globalNotificationPluginInstance,
+        );
+        saveNotificationsUnchecked(const {});
+        break;
+      case AppLifecycleState.inactive:
+        break;
+      case AppLifecycleState.paused:
+        store.dispatch(setBackgrounded(true));
+        break;
+      case AppLifecycleState.detached:
+        store.dispatch(setBackgrounded(true));
+        break;
+    }
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    onMounted();
+  }
+
+  @protected
+  void onMounted() {
+    // init auth listener
+    store.state.authStore.onAuthStateChanged!.listen((user) {
+      if (user == null && defaultHome.runtimeType == HomeScreen) {
+        defaultHome = IntroScreen();
+        NavigationService.clearTo(NavigationPaths.intro, context);
+      } else if (user != null && user.accessToken != null && defaultHome.runtimeType == IntroScreen) {
+        // Default Authenticated App Home
+        defaultHome = HomeScreen();
+        NavigationService.clearTo('/home', context);
+      }
+    });
+
+    // init alerts listener
+    alertsListener = store.state.alertsStore.onAlertsChanged.listen((alert) {
+      Color? color;
+
+      switch (alert.type) {
+        case 'error':
+          color = Colors.red;
+          break;
+        case 'warning':
+          color = Colors.red;
+          break;
+        case 'success':
+          color = Colors.green;
+          break;
+        case 'info':
+        default:
+          color = Colors.grey;
+      }
+
+      final alertMessage = alert.message ?? alert.error ?? 'Unknown Error Occurred';
+
+      globalScaffold.currentState?.showSnackBar(SnackBar(
+        backgroundColor: color,
+        content: Text(
+          alertMessage,
+          style: Theme.of(context).textTheme.subtitle1?.copyWith(color: Colors.white),
+        ),
+        duration: alert.duration,
+        action: SnackBarAction(
+          label: 'Dismiss',
+          textColor: Colors.white,
+          onPressed: () {
+            globalScaffold.currentState?.removeCurrentSnackBar();
+          },
+        ),
+      ));
+    });
+  }
+
+  @override
+  void dispose() {
+    store.dispatch(stopAuthObserver());
+    store.dispatch(stopAlertsObserver());
+    store.dispatch(disposeDeepLinks());
+    closeCache(cache);
+    WidgetsBinding.instance?.removeObserver(this);
+    alertsListener?.cancel();
+    super.dispose();
+  }
+
+  // Store should not need to be passed to a widget to affect
+  // lifecycle widget functions
+  @override
+  Widget build(BuildContext context) => StoreProvider<AppState>(
+        store: store,
+        child: localization.EasyLocalization(
+          path: 'assets/translations',
+          useOnlyLangCode: true,
+          startLocale: Locale(formatLanguageCode(store.state.settingsStore.language)),
+          fallbackLocale: Locale('en'),
+          supportedLocales: const [Locale('en'), Locale('ru'), Locale('pl')],
+          child: StoreConnector<AppState, ThemeSettings>(
+            distinct: true,
+            converter: (store) => store.state.settingsStore.themeSettings,
+            builder: (context, themeSettings) => MaterialApp(
+              localizationsDelegates: context.localizationDelegates,
+              supportedLocales: context.supportedLocales,
+              debugShowCheckedModeBanner: false,
+              theme: setupTheme(themeSettings, generateThemeData: true),
+              navigatorKey: NavigationService.navigatorKey,
+              routes: NavigationProvider.getRoutes(),
+              home: defaultHome,
+              builder: (context, child) => Scaffold(
+                body: child,
+                appBar: null,
+                key: globalScaffold,
+              ),
+            ),
+          ),
+        ),
+      );
+}
